last version of dev + new price algorithm + Order model fix

+++ src/main/java/com/cargohub/models/OrderModel.java
@@ -8,7 +8,7 @@ public class OrderModel {
     private String departureHub;
     private String arrivalHub;
     private Double cargoWeight;
-    private Integer cargoWidth;
-    private Integer cargoHeight;
-    private Integer cargoLength;
+    private Double cargoWidth;
+    private Double cargoHeight;
+    private Double cargoLength;
 }
